refactor(rust): move `traits` module into its own file
