+++ implementations/rust/ockam/ockam_core/src/lib.rs
@@ -51,6 +51,7 @@ mod error;
 mod message;
 mod processor;
 mod routing;
+pub mod traits;
 mod uint;
 pub mod vault;
 mod worker;
@@ -77,27 +78,5 @@ pub mod println_no_std {
         ($($arg:tt)*) => {{
             tracing::info!($($arg)*);
         }};
-    }
-}
-
-/// The traits module provides extended implementations of standard traits.
-pub mod traits {
-    use crate::compat::boxed::Box;
-    use crate::error::Result;
-
-    /// Clone trait for async structs.
-    #[async_trait]
-    pub trait AsyncTryClone: Sized {
-        /// Try cloning a object and return an `Err` in case of failure.
-        async fn async_try_clone(&self) -> Result<Self>;
-    }
-    #[async_trait]
-    impl<D> AsyncTryClone for D
-    where
-        D: Clone + Sync,
-    {
-        async fn async_try_clone(&self) -> Result<Self> {
-            Ok(self.clone())
-        }
     }
 }
